Parallel fixes and 16 bit floats on metal
Finally got around to fixing the bug that caused parallel_inference to not
work on non-Cuda platforms, which in turn had broken interactive_generate
on those platforms.

Also for SDXL on MPS 16 bit floats are working and provide a reasonable
performance boost so enabled them.  This means that something is fucked in
my code for old worlde SD.